Refactoring GPT-sama's spagetti fix for API <> Pydantic <> Model name mismathces
:'( I will plan properly next time, learnt my lesson, I SWEAR!!
